Create review endpoints (#21)
* Start working on reviews

* Create review and review_image migration

* Create list and get endpoints for review start working on sumbmit

* Rename helper.ts to utilities.ts

* Enable strict mode for typescript and refactor everything to match it

* Rename from Globals.ts to globals.ts

* Rename DbResultSet to ResultSet and refactor code

* Create review submit endpoint

* Create update & delete endpoints for review

* Create delete image endpoint for review and refactor CRUD methods to have better response

* Add validation for file extensions

* Update README.md

+++ routes/profile.ts
@@ -1,52 +1,53 @@
 import {UploadedFile} from "express-fileupload";
-import {NextFunction, Request, Response} from "express";
+import {NextFunction, Response} from "express";
+import {Request} from "../helpers/interfaces/request";
 
 export {};
 
 const express = require("express");
 const router = express.Router();
 const userModel = require("../helpers/database/models/user");
-const helper = require("../helpers/helper");
+const utilities = require("../helpers/utilities");
 const translate = require("../helpers/translation");
 const validation = require("../helpers/validation");
 const fs = require("fs");
 const ROLES = require("../helpers/roles");
 
-router.get("/:userID", helper.authenticateToken, async (req: Request, res: Response, _: NextFunction) => {
+router.get("/:userID", utilities.authenticateToken, async (req: Request, res: Response, _: NextFunction) => {
 	const user = await userModel.get(req.params["userID"] || "");
 
-	if (user.rows.length === 0) {
-		return res.status(404).send(helper.invalid_response(translate("account_doesnt_exist")));
+	if (user.data.length === 0) {
+		return res.status(404).send(utilities.invalid_response(translate("account_doesnt_exist")));
 	}
 
 	res.send({
 		"success": user.success,
-		"data": user.rows,
+		"data": user.data,
 		"message": ""
 	});
 });
 
-router.patch("/:userID", helper.authenticateToken, async (req: Request, res: Response, _: NextFunction) => {
-	if (req.params["userID"] !== req["user"]["guid"]) {
-		return res.status(401).send(helper.invalid_response(translate("not_authorized")));
+router.patch("/:userID", utilities.authenticateToken, async (req: Request, res: Response, _: NextFunction) => {
+	if (!req.user || req.params["userID"] !== req["user"]["guid"]) {
+		return res.status(401).send(utilities.invalid_response(translate("not_authorized")));
 	}
 
 	const {name, email} = req.body;
-	const data = {name, email};
+	const data: { name: string, email: string, userID?: number } = {name, email};
 
-	Object.assign(data, {"userID": req.params["userID"]});
+	Object.assign(data, {userID: req.params["userID"]});
 
 	if (req.files && Object.keys(req.files).length > 0) {
 		const image = req.files.image as UploadedFile;
 		const extension = image.name.substring(image.name.lastIndexOf(".") + 1, image.name.length);
 		const imagePath = `/images/user/${data["userID"]}.${extension}`;
 
-		if (fs.existsSync("./public/images/user") === false) {
+		if (!fs.existsSync("./public/images/user")) {
 			fs.mkdirSync("./public/images/user", {recursive: true});
 		}
 
 		const uploadResult = await image.mv(`./public/${imagePath}`).then(() => true).catch(() => false);
-		if (uploadResult === true) {
+		if (uploadResult) {
 			Object.assign(data, {image: imagePath});
 		}
 	}
@@ -57,7 +58,7 @@ router.patch("/:userID", helper.authenticateToken, async (req: Request, res: Res
 	], true);
 
 	if (validationResult.length > 0) {
-		return res.status(400).send(helper.invalid_response(validationResult));
+		return res.status(400).send(utilities.invalid_response(validationResult));
 	}
 
 	const result = await userModel.update(data);
@@ -65,20 +66,20 @@ router.patch("/:userID", helper.authenticateToken, async (req: Request, res: Res
 	res.send({
 		"success": result.success,
 		"data": [],
-		"message": result.success === true ? translate("user_profile_update_success") : translate("unable_to_update_profile")
+		"message": result.success ? translate("user_profile_update_success") : translate("unable_to_update_profile")
 	});
 });
 
-router.delete("/:userID", helper.authenticateToken, async (req: Request, res: Response, _: NextFunction) => {
-	if (req["user"].power < ROLES.Admin && req.params["userID"] !== req["user"]["guid"]) {
-		return res.status(401).send(helper.invalid_response(translate("not_authorized")));
+router.delete("/:userID", utilities.authenticateToken, async (req: Request, res: Response, _: NextFunction) => {
+	if (!req.user || req.user.power < ROLES.Admin && req.params["userID"] !== req["user"]["guid"]) {
+		return res.status(401).send(utilities.invalid_response(translate("not_authorized")));
 	}
 
 	const guid = req.params.userID;
 	const result = await userModel.delete(guid);
 	const success = result.success;
-	const statusCode = success === true ? 200 : 500;
-	const message = success === true ? translate("user_profile_deleted_success") : translate("unable_to_delete_profile");
+	const statusCode = success ? 200 : 500;
+	const message = success ? translate("user_profile_deleted_success") : translate("unable_to_delete_profile");
 
 	res.status(statusCode).send({
 		"success": success,
